% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMRB.R
\name{fetchBMRB}
\alias{fetchBMRB}
\title{This is documentation
This is documentation
This is documentation
This is documentation}
\usage{
fetchBMRB(BMRBidlist)
}
\description{
This is documentation
This is documentation
This is documentation
This is documentation
}

