% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BMRB.R
\name{N15HSQC}
\alias{N15HSQC}
\title{This is documentation
This is documentatiohn
This is documentation}
\usage{
N15HSQC(csdata)
}
\description{
This is documentation
This is documentatiohn
This is documentation
}

